+++ src/Breadcrumb/index.tsx
@@ -1,0 +1,1 @@
+export { Breadcrumb, BreadcrumbCurrentItem, BreadcrumbItem, BreadcrumbMenu } from "./Breadcrumb";