+++ src/Breadcrumbs/index.tsx
@@ -1,1 +1,0 @@
-export { Breadcrumbs, BreadcrumbItem, BreadcrumbMenu } from './Breadcrumbs';